fix: float도 boundary로 지원하도록 수정

+++ beam_anonymizer/ops/anonymizer/grouping.py
@@ -4,7 +4,7 @@ from .base import Anonymizer
 
 
 class Grouping(Anonymizer):
-    def __init__(self, keys: List[str], boundaries: List[int], group_names: List[str]):
+    def __init__(self, keys: List[str], boundaries: List[float], group_names: List[str]):
         super().__init__(keys)
         self.boundaries = boundaries
         self.group_names = group_names
